refactor(app): move seat-the-board nextAction into a shared hook
Keep SeatTheBoard a component-only module for fast refresh. Pages still
route CTAs through SeatTheBoardLink, which reads the same next step.
