+++ app/src/components/index.ts
@@ -3,4 +3,4 @@ export { default as ChamberCard } from './ChamberCard'
 export { default as BoardVisualization } from './BoardVisualization'
 export { default as TreasuryOverview } from './TreasuryOverview'
 export { default as DelegationManager } from './DelegationManager'
-export { default as SeatTheBoard, SeatTheBoardLink, useSeatTheBoard, seatTheBoardHref } from './SeatTheBoard'
+export { default as SeatTheBoard, SeatTheBoardLink } from './SeatTheBoard'
